Feature/cart and offer

+++ src/main/kotlin/org/shangahi/sellio_backend/api/dto/response/OrderHistoryResponse.kt
@@ -1,6 +1,7 @@
 package org.shangahi.sellio_backend.api.dto.response
 
 import org.shangahi.sellio_backend.model.OrderStatus
+import java.math.BigDecimal
 import java.time.Instant
 import java.util.UUID
 
@@ -8,7 +9,7 @@ data class OrderHistoryResponse(
     val orderId: UUID,
     val orderDate: Instant,
     val status: OrderStatus,
-    val totalPrice: Double,
+    val totalPrice: BigDecimal,
     val storeName: String,
     val storeLogoUrl: String?,
     val items: List<OrderItemResponse>
