+++ src/main/kotlin/org/shangahi/sellio_backend/model/TrendingProduct.kt
@@ -1,5 +1,6 @@
 package org.shangahi.sellio_backend.model
 
+import java.math.BigDecimal
 import java.util.UUID
 
 data class TrendingProduct(
@@ -8,6 +9,6 @@ data class TrendingProduct(
     val productDescription: String?,
     val storeId: UUID?,
     val totalSold: Long,
-    val minPrice: Double?,
+    val minPrice: BigDecimal?,
     val mainImageURL: String?
 )
